refactor: App to extends the new BaseApp (#14213)
