+++ packages/astro/src/core/render-context.ts
@@ -42,6 +42,15 @@ export const apiContextRoutesSymbol = Symbol.for('context.routes');
  * Each request is rendered using a `RenderContext`.
  * It contains data unique to each request. It is responsible for executing middleware, calling endpoints, and rendering the page by gathering necessary data from a `Pipeline`.
  */
+
+export type CreateRenderContext = Pick<
+	RenderContext,
+	'pathname' | 'pipeline' | 'request' | 'routeData' | 'clientAddress'
+> &
+	Partial<
+		Pick<RenderContext, 'locals' | 'middleware' | 'status' | 'props' | 'partial' | 'actions'>
+	>;
+
 export class RenderContext {
 	private constructor(
 		readonly pipeline: Pipeline,
@@ -87,10 +96,7 @@ export class RenderContext {
 		props,
 		partial = undefined,
 		actions,
-	}: Pick<RenderContext, 'pathname' | 'pipeline' | 'request' | 'routeData' | 'clientAddress'> &
-		Partial<
-			Pick<RenderContext, 'locals' | 'middleware' | 'status' | 'props' | 'partial' | 'actions'>
-		>): Promise<RenderContext> {
+	}: CreateRenderContext): Promise<RenderContext> {
 		const pipelineMiddleware = await pipeline.getMiddleware();
 		const pipelineActions = actions ?? (await pipeline.getActions());
 		setOriginPathname(
