+++ packages/astro/src/vite-plugin-astro-server/pipeline.ts
@@ -6,6 +6,7 @@ import { enhanceViteSSRError } from '../core/errors/dev/index.js';
 import { AggregateError, CSSError, MarkdownError } from '../core/errors/index.js';
 import type { Logger } from '../core/logger/core.js';
 import type { ModuleLoader } from '../core/module-loader/index.js';
+import { RedirectComponentInstance, routeIsRedirect } from '../core/redirects/index.js';
 import { loadRenderer, Pipeline } from '../core/render/index.js';
 import { createDefaultRoutes } from '../core/routing/default.js';
 import { findRouteToRewrite } from '../core/routing/rewrite.js';
@@ -148,6 +149,10 @@ export class DevPipeline extends Pipeline {
 	}
 
 	async preload(routeData: RouteData, filePath: URL) {
+		if (routeIsRedirect(routeData)) {
+			return RedirectComponentInstance;
+		}
+
 		const { loader } = this;
 
 		// First check built-in routes
