+++ packages/astro/test/actions.test.js
@@ -135,8 +135,9 @@ describe('Astro Actions', () => {
 
 		it('Should fail when calling an action without using Astro.callAction', async () => {
 			const res = await fixture.fetch('/invalid/');
+			assert.equal(res.status, 500);
 			const text = await res.text();
-			assert.match(text, /ActionCalledFromServerError/);
+			assert.match(text, /@vite\/client/);
 		});
 	});
